refactor: extract constant
