+++ test/data-registry.ts
@@ -1,0 +1,3 @@
+import { makeRegistryUrl } from "../src/domain/registry-url";
+
+export const exampleRegistryUrl = makeRegistryUrl("http://example.com");
