+++ test/npmrc-token-update-service.test.ts
@@ -5,10 +5,10 @@ import {
 } from "../src/io/npmrc-io";
 import { AsyncResult, Err, Ok } from "ts-results-es";
 import { RequiredEnvMissingError } from "../src/io/upm-config-io";
-import { exampleRegistryUrl } from "./mock-registry";
 import { tryUpdateUserNpmrcToken } from "../src/services/npmrc-token-update";
 import { IOError } from "../src/common-errors";
 import { emptyNpmrc, setToken } from "../src/domain/npmrc";
+import { exampleRegistryUrl } from "./data-registry";
 
 const exampleNpmrcPath = "/users/someuser/.npmrc";
